requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.7
django-allauth==65.12.0
djangorestframework==3.15.2
django-filter==24.3
idna==3.10
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
pillow==10.4.0
django-cors-headers==4.4.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
django-redis==5.4.0